# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Sofumar Portal API — Auto-generated OpenAPI specification.",
    "description" : "!!! DO NOT EDIT THIS FILE MANUALLY !!!\nThis file is auto-generated by automation pipeline. Any manual changes will be lost during the next build.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:50437",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/users/{id}/status" : {
      "put" : {
        "tags" : [ "User Management" ],
        "summary" : "Toggle user active status",
        "operationId" : "toggleStatus",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserStatusUpdateRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/role" : {
      "put" : {
        "tags" : [ "User Management" ],
        "summary" : "Update user role",
        "operationId" : "updateRole",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRoleUpdateRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/system-settings/update" : {
      "put" : {
        "tags" : [ "System Settings" ],
        "summary" : "Update an existing system setting",
        "operationId" : "updateSystemSetting",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SystemSettingsDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/payments/update" : {
      "put" : {
        "tags" : [ "Payments" ],
        "summary" : "Update an existing payment",
        "operationId" : "updatePayment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/system-settings/search" : {
      "post" : {
        "tags" : [ "System Settings" ],
        "summary" : "Search system settings",
        "operationId" : "searchSystemSettings",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SystemSettingsSearchRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListSystemSettingsDto"
                }
              }
            }
          }
        }
      }
    },
    "/reference/search" : {
      "post" : {
        "tags" : [ "Reference Data" ],
        "summary" : "Search references",
        "operationId" : "searchReferences",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReferenceSearchRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListReferenceDto"
                }
              }
            }
          }
        }
      }
    },
    "/payments/search" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Search payments",
        "operationId" : "searchPayments",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentSearchRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListPaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/payments/add" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Add a new payment",
        "operationId" : "addPayment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseInteger"
                }
              }
            }
          }
        }
      }
    },
...
    "/system-settings/get/{id}" : {
      "get" : {
        "tags" : [ "System Settings" ],
        "summary" : "Get system setting by ID",
        "operationId" : "getSystemSetting",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseSystemSettingsDto"
                }
              }
            }
          }
        }
      }
    },
    "/system-settings/by-key/{key}" : {
      "get" : {
        "tags" : [ "System Settings" ],
        "summary" : "Get settings by key (e.g. Fee, Payment)",
        "operationId" : "getSettingsByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListSystemSettingsDto"
                }
              }
            }
          }
        }
      }
    },
    "/reference/list/{referenceName}" : {
      "get" : {
        "tags" : [ "Reference Data" ],
        "summary" : "Get list of references by name (e.g. feeType)",
        "operationId" : "getReferencesByName",
        "parameters" : [ {
          "name" : "referenceName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListReferenceDescDto"
                }
              }
            }
          }
        }
      }
    },
    "/reference/get/{referenceID}" : {
      "get" : {
        "tags" : [ "Reference Data" ],
        "summary" : "Get reference by ID",
        "operationId" : "getReference",
        "parameters" : [ {
          "name" : "referenceID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseReferenceDto"
                }
              }
            }
          }
        }
      }
    },
    "/payments/latest" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Get latest payments",
        "operationId" : "latestPayments",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 5
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListLatestPaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/payments/get/{paymentID}" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Get payment by ID",
        "operationId" : "getPayment",
        "parameters" : [ {
          "name" : "paymentID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponsePaymentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/payments/delete/{paymentID}" : {
      "delete" : {
        "tags" : [ "Payments" ],
        "summary" : "Delete payment by ID",
        "operationId" : "deletePayment",
        "parameters" : [ {
          "name" : "paymentID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserStatusUpdateRequestDto" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          }
        }
      },
...
      "GlobalMsg" : {
        "type" : "object",
        "description" : "Global-level message",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/MessageType",
            "description" : "Message severity type"
          },
          "message" : {
            "type" : "string",
            "description" : "Message text",
            "example" : "Operation completed successfully."
          }
        }
      },
      "GlobalResponseVoid" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "description" : "Response payload"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
      "UserRoleUpdateRequestDto" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string"
          }
        }
      },
      "SystemSettingsDto" : {
        "type" : "object",
        "properties" : {
          "systemSettingsID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "settingName" : {
            "type" : "string"
          },
          "settingKey" : {
            "type" : "string"
          },
          "settingValue" : {
            "type" : "string"
          },
          "effectiveDate" : {
            "type" : "string",
            "format" : "date"
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "PaymentDto" : {
        "type" : "object",
        "properties" : {
          "paymentID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "memberID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "memberFullName" : {
            "type" : "string"
          },
          "feeType" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "dateReceived" : {
            "type" : "string",
            "format" : "date"
          },
          "methodOfPayment" : {
            "type" : "string"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "quarter" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "SortOrder" : {
        "type" : "string",
        "enum" : [ "asc", "desc" ]
      },
      "SystemSettingsSearchRequestDto" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Page number (0-based)",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "description" : "Page size",
            "example" : 10
          },
          "sortField" : {
            "type" : "string",
            "description" : "Sort field name",
            "example" : "firstName"
          },
          "sortOrder" : {
            "$ref" : "#/components/schemas/SortOrder",
            "description" : "Sort direction",
            "example" : "asc"
          },
          "settingName" : {
            "type" : "string"
          }
        }
      },
      "GlobalResponseListSystemSettingsDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/SystemSettingsDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
      "ReferenceSearchRequestDto" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Page number (0-based)",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "description" : "Page size",
            "example" : 10
          },
          "sortField" : {
            "type" : "string",
            "description" : "Sort field name",
            "example" : "firstName"
          },
          "sortOrder" : {
            "$ref" : "#/components/schemas/SortOrder",
            "description" : "Sort direction",
            "example" : "asc"
          },
          "referenceName" : {
            "type" : "string"
          }
        }
      },
      "GlobalResponseListReferenceDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/ReferenceDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
      "PaymentSearchRequestDto" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Page number (0-based)",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "description" : "Page size",
            "example" : 10
          },
          "sortField" : {
            "type" : "string",
            "description" : "Sort field name",
            "example" : "firstName"
          },
          "sortOrder" : {
            "$ref" : "#/components/schemas/SortOrder",
            "description" : "Sort direction",
            "example" : "asc"
          },
          "memberID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "feeType" : {
            "type" : "string"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "quarter" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dateFrom" : {
            "type" : "string",
            "format" : "date"
          },
          "dateTo" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "GlobalResponseListPaymentDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/PaymentDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
      "GlobalResponseInteger" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Response payload"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}